requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
eventlet==0.29.1
greenlet==0.4.17
idna==2.10
joblib==0.17.0
nltk==3.5
python-engineio==3.13.2
python-socketio==4.6.0
regex==2020.11.13
requests==2.24.0
six==1.15.0
textblob==0.15.3
tqdm==4.53.0
urllib3==1.25.11
websocket-client==0.57.0
